+++ components/LoadingAnimation.js
@@ -1,15 +1,16 @@
-import { StyleSheet, Text, View, Modal, ActivityIndicator} from 'react-native'
-import React from 'react'
+import { StyleSheet, Text, View, Modal} from 'react-native';
+import React from 'react';
 import {PacmanIndicator} from 'react-native-indicators';
-import colors from '../assets/colors/colors'
+import colors from '../assets/colors/colors';
 
-export default LoadingAnimation = () => {
+export default LoadingAnimation = ({caption=''}) => {
   return (
     <Modal transparent={true}
         animationType = 'none'
         >
       <View style={styles.container}>
-        <PacmanIndicator size={48}/>
+        <PacmanIndicator style={{position:'absolute'}} size={48}/>
+        <Text style={{fontFamily:'PixeloidSan', fontSize:15, color:'black', marginBottom:-70}}>{caption}</Text>
       </View>
     </Modal>
   )
